+++ Tests/TestData/SelfVerifyingTests/VBToCS/ObjectEqualityTests.vb
@@ -5,13 +5,13 @@ Imports Xunit
 
 
 Friend Class CTst
+    Public Shared Property GetOperatorResult As Func(Of CTst, CTst, Boolean)
+
     Public Shared Operator <>(ByVal aCls1 As CTst, ByVal aCls2 As CTst) As Boolean
-        Assert.Empty("<> operator should not be called")
-        Return True
+        Return Not GetOperatorResult()(aCls1, aCls2)
     End Operator
     Public Shared Operator =(ByVal aCls1 As CTst, ByVal aCls2 As CTst) As Boolean
-        Assert.Empty("= operator should not be called")
-        Return True
+        Return GetOperatorResult()(aCls1, aCls2)
     End Operator
 End Class
 
@@ -21,6 +21,10 @@ Public Class ObjectEqualityTests
     Public Sub ComparingToNothingDoesNotCallOperatorOverload()
         Dim Test1 As CTst
 
+        CTst.GetOperatorResult = Function(a, b)
+                                     Assert.Empty("operator should not be called")
+                                     Return False
+                                 End Function
         If Not Test1 Is Nothing Then
             Assert.Empty("First branch should not be taken")
         ElseIf Test1 IsNot Nothing Then
@@ -34,6 +38,7 @@ Public Class ObjectEqualityTests
         End If
     End Sub
 
+
     <Fact>
     Public Sub GenericCanBeAssignedNothing()
         Generic(5, False)
@@ -43,5 +48,24 @@ Public Class ObjectEqualityTests
     Private Sub Generic(Of T)(input As T, isReference As Boolean)
         input = Nothing
         If isReference Then Assert.True(input Is Nothing) Else Assert.True(input IsNot Nothing)
+    End Sub
+
+    <Fact>
+    Public Sub ConstrainedGenericCanBeEquatedToNothing()
+        Dim callCount = 0
+        CTst.GetOperatorResult = Function(a, b)
+                                     callCount += 1
+                                     Return Object.ReferenceEquals(a, b)
+                                 End Function
+        GenericEquatable(New CTst)
+        Assert.Equal(callCount, 4)
+    End Sub
+
+    Private Sub GenericEquatable(Of T As CTst)(input As T)
+        Assert.False(input = Nothing)
+        Assert.True(input <> Nothing)
+        input = Nothing
+        Assert.True(input = Nothing)
+        Assert.False(input <> Nothing)
     End Sub
 End Class